file and dir organizations
created src directory and move all source file insiede
